requirements.txt
click>=6.6
sqlalchemy>=1.1
scrapy_splash>=0.7
redis>=2.10
Pillow>=4.0